updated metric calculating

+++ calculate_metrics/calculateCorrelationMatrix.sh
@@ -11,20 +11,22 @@ OUT_FN=$2  # should be a .csv file for the rest of the workflow
 echo $DIR
 echo $OUT_FN
 
-rm $OUT_FN
+if [ -f $OUT_FN ] ; then
+    rm $OUT_FN
+fi
 
 for t in "$DIR"* ; do
     TEMPLATE=$t
-    # echo $TEMPLATE
+    echo $(basename "${TEMPLATE}")
     for f in "$DIR"/* ; do
         # calculate the similarity score and echo it to a file
-        cc=$(./utils/correlation.sh $TEMPLATE $f)
+        cc=$(./correlation.sh $TEMPLATE $f)
         echo -ne "$cc," >> "$OUT_FN"
     done
     # about to have a new template image, want a new line in file
     echo -e "" >> "$OUT_FN"
 done
 
-python utils/processCorrelationFile.py -f $OUT_FN
+python processCorrelationFile.py -f $OUT_FN
 
 echo "Finished calculating correlation matrix."
